+++ src/libserver/http/http_router.c
@@ -291,7 +291,7 @@ rspamd_http_router_finish_handler (struct rspamd_http_connection *conn,
 			http_parser_parse_url (msg->url->str, msg->url->len, TRUE, &u);
 
 			if (u.field_set & (1 << UF_PATH)) {
-				guint unnorm_len;
+				gsize unnorm_len;
 
 				pathbuf = g_malloc (u.field_data[UF_PATH].len);
 				memcpy (pathbuf, msg->url->str + u.field_data[UF_PATH].off,
